+++ modules/secure-cloud-run-security/versions.tf
@@ -29,10 +29,10 @@ terraform {
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-cloud-run:secure-cloud-run-security/v0.7.0"
+    module_name = "blueprints/terraform/terraform-google-cloud-run:secure-cloud-run-security/v0.9.1"
   }
 
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-cloud-run:secure-cloud-run-security/v0.7.0"
+    module_name = "blueprints/terraform/terraform-google-cloud-run:secure-cloud-run-security/v0.9.1"
   }
 }