+++ modules/secure-serverless-harness/main.tf
@@ -15,8 +15,7 @@
  */
 
 locals {
-  api = var.serverless_type == "CLOUD_FUNCTION" ? ["cloudfunctions.googleapis.com", "cloudbuild.googleapis.com", "eventarc.googleapis.com", "eventarcpublishing.googleapis.com"] : []
-  serverless_apis = concat([
+  serverless_apis = [
     "vpcaccess.googleapis.com",
     "compute.googleapis.com",
     "container.googleapis.com",
@@ -25,7 +24,7 @@ locals {
     "cloudkms.googleapis.com",
     "dns.googleapis.com",
     "servicenetworking.googleapis.com"
-  ], local.api)
+  ]
   kms_apis = concat([
     "cloudkms.googleapis.com",
     "artifactregistry.googleapis.com"
@@ -60,6 +59,8 @@ module "network_project" {
   billing_account   = var.billing_account
   folder_id         = google_folder.fld_serverless.name
 
+  disable_services_on_destroy = var.disable_services_on_destroy
+
   enable_shared_vpc_host_project = true
 }
 
@@ -72,6 +73,8 @@ module "security_project" {
   org_id            = var.org_id
   billing_account   = var.billing_account
   folder_id         = google_folder.fld_serverless.name
+
+  disable_services_on_destroy = var.disable_services_on_destroy
 }
 
 module "serverless_project" {
@@ -80,16 +83,19 @@ module "serverless_project" {
   for_each = toset(var.serverless_project_names)
 
   billing_account               = var.billing_account
-  serverless_type               = var.serverless_type
+  base_serverless_api           = var.base_serverless_api
   org_id                        = var.org_id
   activate_apis                 = concat(local.serverless_apis, try(var.serverless_project_extra_apis[each.value], []))
   folder_name                   = google_folder.fld_serverless.name
   project_name                  = each.value
   service_account_project_roles = try(var.service_account_project_roles[each.value], [])
+
+  disable_services_on_destroy = var.disable_services_on_destroy
 }
 
 
 resource "google_artifact_registry_repository" "repo" {
+  count         = var.base_serverless_api == "run.googleapis.com" ? 1 : 0
   project       = module.security_project.project_id
   location      = var.location
   repository_id = var.artifact_registry_repository_name
@@ -98,20 +104,20 @@ resource "google_artifact_registry_repository" "repo" {
   kms_key_name  = module.artifact_registry_kms.keys[var.key_name]
 
   depends_on = [
-    time_sleep.wait_180_seconds
+    time_sleep.wait_vpc_sc_propagation
   ]
 }
 
 resource "google_artifact_registry_repository_iam_member" "member" {
-  for_each   = module.serverless_project
+  for_each   = var.base_serverless_api == "run.googleapis.com" ? module.serverless_project : {}
   project    = module.security_project.project_id
   location   = var.location
-  repository = google_artifact_registry_repository.repo.repository_id
+  repository = google_artifact_registry_repository.repo[0].repository_id
   role       = "roles/artifactregistry.reader"
   member     = "serviceAccount:${each.value.cloud_serverless_service_identity_email}"
 
   depends_on = [
-    time_sleep.wait_180_seconds
+    time_sleep.wait_vpc_sc_propagation
   ]
 }
 
@@ -134,7 +140,7 @@ module "artifact_registry_kms" {
   key_protection_level = var.key_protection_level
 
   depends_on = [
-    time_sleep.wait_180_seconds
+    time_sleep.wait_vpc_sc_propagation
   ]
 }
 
@@ -145,27 +151,6 @@ resource "google_project_service_identity" "artifact_sa" {
   service = "artifactregistry.googleapis.com"
 
   depends_on = [
-    time_sleep.wait_180_seconds
-  ]
-}
-
-module "cloudfunction_source_bucket" {
-  for_each = var.serverless_type == "CLOUD_RUN" ? {} : module.serverless_project
-  source   = "terraform-google-modules/cloud-storage/google//modules/simple_bucket"
-  version  = "~>3.4"
-
-  project_id    = each.value.project_id
-  name          = "bkt-${var.location}-${each.value.project_number}-cfv2-zip-files"
-  location      = var.location
-  storage_class = "REGIONAL"
-  force_destroy = true
-
-  encryption = {
-    default_kms_key_name = module.artifact_registry_kms.keys[var.key_name]
-  }
-
-  depends_on = [
-    module.artifact_registry_kms,
-    time_sleep.wait_180_seconds
+    time_sleep.wait_vpc_sc_propagation
   ]
 }