+++ modules/secure-serverless-harness/outputs.tf
@@ -19,7 +19,7 @@ output "serverless_folder_id" {
   description = "The folder created to alocate Serverless infra."
 
   depends_on = [
-    time_sleep.wait_180_seconds
+    time_sleep.wait_vpc_sc_propagation
   ]
 }
 
@@ -28,7 +28,7 @@ output "network_project_id" {
   description = "Project ID of the project created to host the Cloud Run Network."
 
   depends_on = [
-    time_sleep.wait_180_seconds
+    time_sleep.wait_vpc_sc_propagation
   ]
 }
 
@@ -37,7 +37,7 @@ output "serverless_project_ids" {
   description = "Project ID of the projects created to deploy Cloud Run."
 
   depends_on = [
-    time_sleep.wait_180_seconds
+    time_sleep.wait_vpc_sc_propagation
   ]
 }
 
@@ -46,7 +46,7 @@ output "serverless_project_numbers" {
   description = "Project number of the projects created to deploy Cloud Run."
 
   depends_on = [
-    time_sleep.wait_180_seconds
+    time_sleep.wait_vpc_sc_propagation
   ]
 }
 
@@ -55,7 +55,7 @@ output "security_project_id" {
   description = "Project ID of the project created for KMS and Artifact Register."
 
   depends_on = [
-    time_sleep.wait_180_seconds
+    time_sleep.wait_vpc_sc_propagation
   ]
 }
 
@@ -64,7 +64,7 @@ output "security_project_number" {
   description = "Project number of the project created for KMS and Artifact Register."
 
   depends_on = [
-    time_sleep.wait_180_seconds
+    time_sleep.wait_vpc_sc_propagation
   ]
 }
 
@@ -73,7 +73,7 @@ output "service_account_email" {
   description = "The email of the Service Account created to be used by Cloud Serverless."
 
   depends_on = [
-    time_sleep.wait_180_seconds
+    time_sleep.wait_vpc_sc_propagation
   ]
 }
 
@@ -82,7 +82,7 @@ output "service_vpc" {
   description = "The network created for Cloud Serverless."
 
   depends_on = [
-    time_sleep.wait_180_seconds
+    time_sleep.wait_vpc_sc_propagation
   ]
 }
 
@@ -91,25 +91,25 @@ output "service_subnet" {
   description = "The sub-network name created in harness."
 
   depends_on = [
-    time_sleep.wait_180_seconds
+    time_sleep.wait_vpc_sc_propagation
   ]
 }
 
 output "artifact_registry_repository_id" {
-  value       = google_artifact_registry_repository.repo.id
+  value       = var.base_serverless_api == "run.googleapis.com" ? google_artifact_registry_repository.repo[0].id : ""
   description = "The Artifact Registry Repository full identifier where the images should be stored."
 
   depends_on = [
-    time_sleep.wait_180_seconds
+    time_sleep.wait_vpc_sc_propagation
   ]
 }
 
 output "artifact_registry_repository_name" {
-  value       = google_artifact_registry_repository.repo.repository_id
+  value       = var.base_serverless_api == "run.googleapis.com" ? google_artifact_registry_repository.repo[0].repository_id : ""
   description = "The Artifact Registry Repository last part of the repository name where the images should be stored."
 
   depends_on = [
-    time_sleep.wait_180_seconds
+    time_sleep.wait_vpc_sc_propagation
   ]
 }
 
@@ -118,7 +118,16 @@ output "cloud_serverless_service_identity_email" {
   description = "The Cloud Run Service Identity email."
 
   depends_on = [
-    time_sleep.wait_180_seconds
+    time_sleep.wait_vpc_sc_propagation
+  ]
+}
+
+output "access_context_manager_policy_id" {
+  value       = local.access_context_manager_policy_id
+  description = "Access Context Manager ID."
+
+  depends_on = [
+    time_sleep.wait_vpc_sc_propagation
   ]
 }
 
@@ -127,7 +136,7 @@ output "restricted_service_perimeter_name" {
   description = "Service Perimeter name."
 
   depends_on = [
-    time_sleep.wait_180_seconds
+    time_sleep.wait_vpc_sc_propagation
   ]
 }
 
@@ -136,15 +145,24 @@ output "restricted_access_level_name" {
   description = "Access level name."
 
   depends_on = [
-    time_sleep.wait_180_seconds
+    time_sleep.wait_vpc_sc_propagation
   ]
 }
 
-output "cloudfunction_source_bucket" {
-  value       = var.serverless_type == "CLOUD_RUN" ? {} : { for bucket in module.cloudfunction_source_bucket : bucket.bucket.project => bucket.bucket }
-  description = "Cloud Function Source Bucket."
+output "restricted_access_level_name_id" {
+  value       = module.access_level_members.name_id
+  description = "Access level name id."
 
   depends_on = [
-    time_sleep.wait_180_seconds
+    time_sleep.wait_vpc_sc_propagation
+  ]
+}
+
+output "artifact_registry_key" {
+  value       = module.artifact_registry_kms.keys[var.key_name]
+  description = "Artifact Registry KMS Key."
+
+  depends_on = [
+    time_sleep.wait_vpc_sc_propagation
   ]
 }