+++ modules/secure-serverless-harness/service_perimeter.tf
@@ -21,7 +21,8 @@ locals {
   access_context_manager_policy_id = var.create_access_context_manager_access_policy ? google_access_context_manager_access_policy.access_policy[0].id : var.access_context_manager_policy_id
   access_level_members = concat(var.access_level_members,
     [for project in module.serverless_project : "serviceAccount:${project.services_identities["cloudbuild"]}"],
-    [for project in module.serverless_project : "serviceAccount:${project.services_identities["gcs"]}"]
+    [for project in module.serverless_project : "serviceAccount:${project.services_identities["gcs"]}"],
+    [for project in module.serverless_project : "serviceAccount:${project.services_identities["cloudservices"]}"]
   )
 }
 
@@ -220,14 +221,15 @@ resource "google_access_context_manager_service_perimeter_resource" "service_per
   ]
 }
 
-resource "time_sleep" "wait_180_seconds" {
+resource "time_sleep" "wait_vpc_sc_propagation" {
   depends_on = [
     google_access_context_manager_service_perimeter_resource.service_perimeter_security_resource,
     google_access_context_manager_service_perimeter_resource.service_perimeter_serverless_resource,
     google_access_context_manager_service_perimeter_resource.service_perimeter_network_resource,
-    module.access_level_members
+    module.access_level_members,
+    module.regular_service_perimeter
   ]
 
-  create_duration  = "180s"
-  destroy_duration = "180s"
+  create_duration  = var.time_to_wait_vpc_sc_propagation
+  destroy_duration = var.time_to_wait_vpc_sc_propagation
 }