+++ modules/secure-serverless-harness/versions.tf
@@ -37,10 +37,10 @@ terraform {
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-cloud-run:secure-serverless-harness/v0.7.0"
+    module_name = "blueprints/terraform/terraform-google-cloud-run:secure-serverless-harness/v0.9.1"
   }
 
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-cloud-run:secure-serverless-harness/v0.7.0"
+    module_name = "blueprints/terraform/terraform-google-cloud-run:secure-serverless-harness/v0.9.1"
   }
 }