+++ modules/service-project-factory/main.tf
@@ -14,19 +14,16 @@
  * limitations under the License.
  */
 
-
-locals {
-  api = var.serverless_type == "CLOUD_RUN" ? "run" : "cloudfunctions"
-}
 module "serverless_project" {
-  source            = "terraform-google-modules/project-factory/google"
-  version           = "~> 14.2"
-  random_project_id = "true"
-  activate_apis     = var.activate_apis
-  name              = var.project_name
-  org_id            = var.org_id
-  billing_account   = var.billing_account
-  folder_id         = var.folder_name
+  source                      = "terraform-google-modules/project-factory/google"
+  version                     = "~> 14.2"
+  random_project_id           = "true"
+  activate_apis               = var.activate_apis
+  name                        = var.project_name
+  org_id                      = var.org_id
+  billing_account             = var.billing_account
+  folder_id                   = var.folder_name
+  disable_services_on_destroy = var.disable_services_on_destroy
 
   svpc_host_project_id = var.network_project_id
   grant_network_role   = var.network_project_id != "" ? true : false
@@ -37,7 +34,7 @@ module "service_accounts" {
   version    = "~> 4.2"
   project_id = module.serverless_project.project_id
   prefix     = "sa"
-  names      = [replace(lower(var.serverless_type), "_", "-")]
+  names      = var.base_serverless_api == "run.googleapis.com" ? ["cloud-run"] : ["cloud-function"]
 
   depends_on = [
     module.serverless_project
@@ -55,7 +52,7 @@ resource "google_project_service_identity" "serverless_sa" {
   provider = google-beta
 
   project = module.serverless_project.project_id
-  service = "${local.api}.googleapis.com"
+  service = var.base_serverless_api
 }
 
 resource "google_service_account_iam_member" "identity_service_account_user" {
@@ -83,7 +80,7 @@ data "google_storage_project_service_account" "gcs_account" {
 }
 
 resource "google_project_iam_member" "gcs_pubsub_publishing" {
-  count   = var.serverless_type == "CLOUD_RUN" ? 0 : 1
+  count   = var.base_serverless_api == "run.googleapis.com" ? 0 : 1
   project = module.serverless_project.project_id
   role    = "roles/pubsub.publisher"
   member  = "serviceAccount:${data.google_storage_project_service_account.gcs_account.email_address}"