+++ test/integration/secure_cloud_run_standalone/secure_cloud_run_standalone_test.go
@@ -304,9 +304,9 @@ func TestSecureCloudRunStandalone(t *testing.T) {
 			assert.Equal(connectorId, opVPCConnector.Get("name").String(), fmt.Sprintf("Should have same id: %s", connectorId))
 			assert.Equal(expectedSubnet, opVPCConnector.Get("subnet.name").String(), fmt.Sprintf("Should have same subnetwork: %s", expectedSubnet))
 			assert.Equal(expectedMachineType, opVPCConnector.Get("machineType").String(), fmt.Sprintf("Should have same machineType: %s", expectedMachineType))
-			assert.Equal("7", opVPCConnector.Get("maxInstances").String(), "Should have maxInstances equals to 7")
+			assert.Equal("10", opVPCConnector.Get("maxInstances").String(), "Should have maxInstances equals to 10")
 			assert.Equal("2", opVPCConnector.Get("minInstances").String(), "Should have minInstances equals to 2")
-			assert.Equal("700", opVPCConnector.Get("maxThroughput").String(), "Should have maxThroughput equals to 700")
+			assert.Equal("1000", opVPCConnector.Get("maxThroughput").String(), "Should have maxThroughput equals to 1000")
 			assert.Equal("200", opVPCConnector.Get("minThroughput").String(), "Should have minThroughput equals to 200")
 
 			// Cloud Armor tests